Connecting and Integrating the TiffDataReader class into OpenHKL In this comment several things have been done outside the new class. Biggest change is the new method Session::loadTiffData() which is called when a *.tif file will be imported to the program. A new DataFormat key DataFormat::TIF has been introduced to destinguish between tif and other formats as well. Additionaly a new meta data flag has been introduced to store and retrieve the file resolution of each imported tif file.

+++ core/data/DataSet.cpp
@@ -47,6 +47,7 @@
 #include <regex>
 #include <regex>
 #include <stdexcept>
+#include <regex>
 
 namespace ohkl {
 
@@ -223,12 +224,12 @@ std::size_t DataSet::nFrames() const
 
 std::size_t DataSet::nCols() const
 {
-    return detector().nCols();
+    return (float)detector().nCols();
 }
 
 std::size_t DataSet::nRows() const
 {
-    return detector().nRows();
+    return (float)detector().nRows();
 }
 
 double DataSet::wavelength() const
